Add files via upload
The system includes a terminal fingerprint acquisition module and an attendance module. It can realize automatically such functions as information acquisition of fingerprints, processing, and wireless transmission, fingerprint matching, and making an attendance report. After taking the attendance, this system sends the attendance of every student to their own mobile and their parent’s mobile through the API system and also stored the attendance of respective students to calculate the percentage of attendance and alerts to class in charge. The attendance system facilitates access to the attendance of a particular student in a particular class. This system eliminates the need for stationary materials and personnel for the keeping of records and efforts of the class in charge.